agent-backend/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["google-genai>=1.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
